refactor: updated to activity log version 1
