+++ src/App.tsx
@@ -25,7 +25,7 @@ import { InventoryDetails } from './pages/inventory/inventory-details';
 import { SidebarProvider } from 'components/ui/sidebar';
 import { Email } from './pages/email/email';
 import { VerifyOtpKey } from './pages/auth/verify-otp-key/verify-otp-key';
-import ActivityLog from './pages/activity-log/activity-log';
+import ActivityLog from './pages/activity-log-v1/activity-log';
 import { InventoryForm } from './features/inventory/component/inventory-form/inventory-form';
 
 const queryClient = new QueryClient();
@@ -73,7 +73,7 @@ function AppContent() {
                 <Route path="/inventory/add" element={<InventoryForm />} />
                 <Route path="/inventory/:itemId" element={<InventoryDetails />} />
                 <Route path="/mail" element={<Email />} />
-                <Route path="/activity-log" element={<ActivityLog />} />
+                <Route path="/activity-log-v1" element={<ActivityLog />} />
                 <Route path="/help" element={<Help />} />
                 <Route path="/identity-management" element={<TaskPage />} />
                 <Route path="/services/storage" element={<Storage />} />
