+++ src/constant/dynamic-breadcrumb-title.ts
@@ -5,6 +5,6 @@ export const DYNAMIC_BREADCRUMB_TITLES: Record<string, string | null> = {
   '/inventory/[id]': 'Inventory',
   '/profile': 'Profile',
   '/mail': 'Mail',
-  '/activity-log': 'Activity log',
+  '/activity-log-v1': 'Activity log',
   '/inventory/add': 'Add item',
 };
